% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cran_pkg_graph}
\alias{cran_pkg_graph}
\title{Package-Author tidygraph from a CRAN database snapshot}
\format{A large tbl_graph (16911 nodes, 99542 edges, 10 elements, 5.8 Mb)
\describe{
  \item{Nodes}{package authors}
  \item{Edges}{package names}
}}
\source{
\url{https://cran.r-project.org/}
}
\usage{
data(cran_pkg_data)
}
\description{
A tidygraph object of all package-author relationships as captured on 26
February 2019. Extracted using the \code{kevinbacran::kb_combos()} function.
}
\keyword{datasets}
