% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_and_prep.R
\name{kb_combos}
\alias{kb_combos}
\title{Fetch CRAN Authors And Create Tidygraph}
\usage{
kb_combos()
}
\value{
A tidygraph object with classes 'tbl_graph' and 'igraph'.
}
\description{
Fetches CRAN package data, cleans author names, creates a tidy
    tibble of each package-author combination and turns this into a tidygraph
    object. Nodes are authors and edges are packages they collaborate on.
}
\examples{
\dontrun{
cran_graph <- kb_combos()
print(cran_graph)
}
}
