% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_and_prep.R
\name{kb_pair}
\alias{kb_pair}
\title{Get Author-Package Network For Two Named Authors}
\usage{
kb_pair(tidy_graph, name_a, name_b = "Hadley Wickham")
}
\arguments{
\item{tidy_graph}{A tidygraph object of CRAN authors/packages created with
\code{kb_combos()}}

\item{name_a}{A character string of a CRAN author's name (one of two)}

\item{name_b}{A character string of a CRAN author's name (two of two)}
}
\value{
A tidygraph object with classes 'tbl_graph' and 'igraph'.
}
\description{
Finds one of the possible shortest paths between two named
    authors from CRAN. This is possible by searching the tidygraph object of
    all CRAN's package-author relationships created with
    \code{kevinbacran::kb_combos()}. Outputs a tidygraph object that's a
    subset of the full CRAN tidygraph. This output contains only the named
    authors and the authors and packages that link them together.
}
\examples{
\dontrun{
pair_graph <- kb_pair(cran_graph, "Garrett Grolemund", "Yihui Xie")
print(pair_graph)
}
}
