% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_outputs.R
\name{kb_plot}
\alias{kb_plot}
\title{Simple Plot Of Shortest Distance Between CRAN Authors}
\usage{
kb_plot(pair_graph)
}
\arguments{
\item{pair_graph}{A tidygraph object of CRAN authors/packages created with
\code{kb_pair()}}
}
\value{
A ggraph plot
}
\description{
Returns a ggplot2 object that represents the degrees of
    separation between the two named authors that were provided to
    \code{kevinbacran::kb_pair()}. The two named authors are terminal nodes.
    Intermediate nodes represent intermediate authors that link via edges
    that represent a package that the authors collaborated on.
}
\examples{
\dontrun{
 <- kb_plot(pair_graph)
print(separation)
}
}
