% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kevinbacran-package.R
\docType{package}
\name{kevinbacran-package}
\alias{kevinbacran}
\alias{kevinbacran-package}
\title{kevinbacran: Degrees of Separation Between CRAN authors}
\description{
Calculate the shortest path through a graph network between any two 
    authors listed on CRAN. Analagous to the Six Degrees of Kevin Bacon or Erdos
    Number. Author X collaborates with author Y; author Z collaborates with Y
    but not X, therefore the distance from X to Z is two.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matt-dray/kevinbacran}
  \item Report bugs at \url{https://github.com/matt-dray/kevinbacran/issues}
}

}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
